sign with libp2p private key
