+++ packages/network/src/node.ts
@@ -17,6 +17,7 @@ import type {
 	Address,
 	EventCallback,
 	PeerDiscovery,
+	Secp256k1PrivateKey,
 	Stream,
 	StreamHandler,
 } from "@libp2p/interface";
@@ -30,9 +31,10 @@ import { webSockets } from "@libp2p/websockets";
 import * as filters from "@libp2p/websockets/filters";
 import { type MultiaddrInput, multiaddr } from "@multiformats/multiaddr";
 import { WebRTC } from "@multiformats/multiaddr-matcher";
-import { etc } from "@noble/secp256k1";
+import { etc, signAsync } from "@noble/secp256k1";
 import { Logger, type LoggerOptions } from "@ts-drp/logger";
 import { Message } from "@ts-drp/types";
+import * as crypto from "crypto";
 import { type Libp2p, type ServiceFactoryMap, createLibp2p } from "libp2p";
 import { fromString as uint8ArrayFromString } from "uint8arrays/from-string";
 
@@ -69,6 +71,7 @@ export class DRPNetworkNode {
 	private _config?: DRPNetworkNodeConfig;
 	private _node?: Libp2p;
 	private _pubsub?: GossipSub;
+	private _secp256k1PrivateKey?: Secp256k1PrivateKey;
 
 	peerId = "";
 
@@ -80,15 +83,14 @@ export class DRPNetworkNode {
 	async start() {
 		if (this._node?.status === "started") throw new Error("Node already started");
 
-		let privateKey = undefined;
 		if (this._config?.private_key_seed) {
 			const tmp = this._config.private_key_seed.padEnd(64, "0");
 			const seed = uint8ArrayFromString(tmp);
 			const seedHex = etc.bytesToHex(seed);
 			const rawPrivateKey = etc.hashToPrivateKey(seedHex);
-			privateKey = privateKeyFromRaw(rawPrivateKey);
+			this._secp256k1PrivateKey = privateKeyFromRaw(rawPrivateKey) as Secp256k1PrivateKey;
 		} else {
-			privateKey = await generateKeyPair("secp256k1");
+			this._secp256k1PrivateKey = await generateKeyPair("secp256k1");
 		}
 
 		const _bootstrapNodesList = this._config?.bootstrap_peers
@@ -174,7 +176,7 @@ export class DRPNetworkNode {
 		};
 
 		this._node = await createLibp2p({
-			privateKey,
+			privateKey: this._secp256k1PrivateKey,
 			addresses: {
 				listen: this._config?.listen_addresses
 					? this._config.listen_addresses
@@ -414,5 +416,24 @@ export class DRPNetworkNode {
 
 	async addCustomMessageHandler(protocol: string | string[], handler: StreamHandler) {
 		await this._node?.handle(protocol, handler);
+	}
+
+	async signWithSecp256k1(data: string): Promise<Uint8Array> {
+		if (!this._secp256k1PrivateKey) {
+			throw new Error("Private key not found");
+		}
+		const hashData = crypto.createHash("sha256").update(data).digest("hex");
+
+		const signature = await signAsync(hashData, this._secp256k1PrivateKey.raw, {
+			extraEntropy: true,
+		});
+
+		const compactSignature = signature.toCompactRawBytes();
+
+		const fullSignature = new Uint8Array(1 + compactSignature.length);
+		fullSignature[0] = signature.recovery;
+		fullSignature.set(compactSignature, 1);
+
+		return fullSignature;
 	}
 }
