+++ packages/node/src/handlers.ts
@@ -372,7 +372,7 @@ export async function signGeneratedVertices(node: DRPNode, vertices: Vertex[]) {
 			return;
 		}
 		try {
-			vertex.signature = await node.credentialStore.signWithSecp256k1(vertex.hash);
+			vertex.signature = await node.networkNode.signWithSecp256k1(vertex.hash);
 		} catch (error) {
 			log.error("::signGeneratedVertices: Error signing vertex:", vertex.hash, error);
 		}
